requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.0
google-auth>=2.16.0
python-dotenv>=1.0.0
langdetect>=1.0.9
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
